Moved custom regex validation constraint ns
